refactor(shared): move internal/github → shared/github
