+++ cmd/agentdock/app.go
@@ -12,7 +12,7 @@ import (
 	"github.com/Ivantseng123/agentdock/internal/bot"
 	"github.com/Ivantseng123/agentdock/internal/config"
 	"github.com/Ivantseng123/agentdock/shared/crypto"
-	ghclient "github.com/Ivantseng123/agentdock/internal/github"
+	ghclient "github.com/Ivantseng123/agentdock/shared/github"
 	"github.com/Ivantseng123/agentdock/shared/logging"
 	"github.com/Ivantseng123/agentdock/internal/mantis"
 	"github.com/Ivantseng123/agentdock/shared/metrics"
