+++ shared/github/discovery.go
